+++ src/display/components/screenrotator.tsx
@@ -294,6 +294,16 @@ export function ScreenRotator(props: ScreenRotatorProps) {
           />
         );
 
+      case Screens.CANADA_TEMP_CONDITIONS_ON:
+        return (
+          <NationalWeatherScreen
+            weatherStationTime={weatherStationResponse?.stationTime}
+            observations={nationalWeather?.on}
+            area="ON"
+            onComplete={switchToNextScreen}
+          />
+        );
+
       case Screens.CANADA_TEMP_CONDITIONS_WEST:
         return (
           <NationalWeatherScreen